+++ connectors/object/s3/config.go
