+++ connectors/object/s3/crc.go
